third_party/blink/web_tests/accessibility/scroll-window-sends-notification.html

<!DOCTYPE HTML>
<html>
<head>
  <script src="../resources/testharness.js"></script>
  <script src="../resources/testharnessreport.js"></script>
  <script src="../resources/run-after-layout-and-paint.js"></script>
<style>
.bigbutton {
    display:block;
    width: 600px;
    height: 600px;
}
</style>
</head>
<body>

<button class="bigbutton">One</button>
<button class="bigbutton">Two</button>
<button class="bigbutton">Three</button>

<div id="console"></div>
<script>

async_test_after_layout_and_paint((t) => {
    window.scrollTo(0, 0);
    assert_equals(window.pageYOffset, 0);

    accessibilityController.addNotificationListener(function (target, notification) {
        if (target.role == 'AXRole: AXWebArea' && notification == 'ScrollPositionChanged') {
            console.log('Got notification on web area');
            assert_equals(window.pageYOffset, 500);
            t.done()
        }
    });

    window.setTimeout(function() {
        window.scrollTo(0, 500);
    }, 0);
}, "This test ensures that scrolling the window sends a notification.");


</script>

</body>
</html>
